move manual compact to file
